[NEW RBO] Refactored optimizer rules
