+++ ydb/core/kqp/opt/rbo/rules/build_initial_cbo_tree.cpp
@@ -1,0 +1,32 @@
+#include "kqp_rules_include.h"
+
+namespace NKikimr {
+namespace NKqp {
+
+/**
+ * Initially we build CBO only for joins that don't have other joins or CBO trees as arguments
+ * There could be an intermediate filter in between, we also check that
+ */
+std::shared_ptr<IOperator> TBuildInitialCBOTreeRule::SimpleMatchAndApply(const std::shared_ptr<IOperator> &input, TRBOContext &ctx, TPlanProps &props) {
+    Y_UNUSED(ctx);
+    Y_UNUSED(props);
+
+    auto containsJoins = [](const std::shared_ptr<IOperator>& op) {
+        std::shared_ptr<IOperator> maybeJoin = op;
+        if (op->Kind == EOperator::Filter) {
+            maybeJoin = CastOperator<TOpFilter>(op)->GetInput();
+        }
+        return (maybeJoin->Kind == EOperator::Join || maybeJoin->Kind == EOperator::CBOTree);
+    };
+
+    if (input->Kind == EOperator::Join) {
+        auto join = CastOperator<TOpJoin>(input);
+        if (!containsJoins(join->GetLeftInput()) && !containsJoins(join->GetRightInput())) {
+            return std::make_shared<TOpCBOTree>(input, input->Pos);
+        }
+    }
+
+    return input;
+}
+}
+}
