+++ ydb/core/kqp/opt/rbo/rules/kqp_rules_include.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <ydb/core/kqp/opt/rbo/kqp_rbo_rules.h>
+#include <ydb/core/kqp/opt/rbo/kqp_operator.h>
+#include <ydb/core/kqp/opt/rbo/kqp_rbo_utils.h>
+#include <ydb/core/kqp/common/kqp_yql.h>
+#include <yql/essentials/core/yql_expr_optimize.h>
+#include <yql/essentials/utils/log/log.h>
+#include <ydb/core/kqp/opt/physical/predicate_collector.h>
+#include <ydb/core/kqp/opt/physical/kqp_opt_phy_olap_filter.h>
+#include <yql/essentials/core/peephole_opt/yql_opt_peephole_physical.h>
+
+#include <typeinfo>
