+++ ydb/core/kqp/opt/rbo/rules/push_limit_into_sort.cpp
@@ -1,0 +1,28 @@
+#include "kqp_rules_include.h"
+
+namespace NKikimr {
+namespace NKqp {
+    
+std::shared_ptr<IOperator> TPushLimitIntoSortRule::SimpleMatchAndApply(const std::shared_ptr<IOperator> &input, TRBOContext &ctx, TPlanProps &props) {
+    Y_UNUSED(ctx);
+    Y_UNUSED(props);
+
+    if (input->Kind != EOperator::Limit) {
+        return input;
+    }
+
+    auto limit = CastOperator<TOpLimit>(input);
+    if (limit->GetInput()->Kind != EOperator::Sort) {
+        return input;
+    }
+
+    auto sort = CastOperator<TOpSort>(limit->GetInput());
+    if (sort->LimitCond) {
+        return input;
+    }
+
+    sort->LimitCond = limit->LimitCond;
+    return sort;
+}
+}
+}
