+++ ydb/core/kqp/opt/rbo/rules/push_olap_filter.cpp
@@ -1,0 +1,101 @@
+#include "kqp_rules_include.h"
+
+
+namespace {
+using namespace NYql::NNodes;
+using namespace NKikimr;
+using namespace NKikimr::NKqp;
+
+bool IsSuitableToPushPredicateToColumnTables(const std::shared_ptr<IOperator>& input) {
+    if (input->Kind != EOperator::Filter) {
+        return false;
+    }
+
+    const auto filter = CastOperator<TOpFilter>(input);
+    const auto maybeRead = filter->GetInput();
+    return ((maybeRead->Kind == EOperator::Source) && (CastOperator<TOpRead>(maybeRead)->GetTableStorageType() == NYql::EStorageType::ColumnStorage) &&
+            filter->GetTypeAnn());
+}
+}
+
+namespace NKikimr {
+namespace NKqp {
+
+std::shared_ptr<IOperator> TPushOlapFilterRule::SimpleMatchAndApply(const std::shared_ptr<IOperator>& input, TRBOContext& ctx, TPlanProps& props) {
+    Y_UNUSED(props);
+    if (!ctx.KqpCtx.Config->HasOptEnableOlapPushdown()) {
+        return input;
+    }
+
+    const TPushdownOptions pushdownOptions(ctx.KqpCtx.Config->GetEnableOlapScalarApply(), ctx.KqpCtx.Config->GetEnableOlapSubstringPushdown(),
+                                           /*StripAliasPrefixForColumnName=*/true);
+    if (!IsSuitableToPushPredicateToColumnTables(input)) {
+        return input;
+    }
+
+    const auto filter = CastOperator<TOpFilter>(input);
+    const auto read = CastOperator<TOpRead>(filter->GetInput());
+    const auto lambda = TCoLambda(filter->FilterExpr.Node);
+    const auto& lambdaArg = lambda.Args().Arg(0).Ref();
+    TExprBase predicate = lambda.Body();
+
+    TOLAPPredicateNode predicateTree;
+    predicateTree.ExprNode = predicate.Ptr();
+    CollectPredicates(predicate, predicateTree, &lambdaArg, filter->GetTypeAnn()->Cast<TListExprType>()->GetItemType(), pushdownOptions);
+    YQL_ENSURE(predicateTree.IsValid(), "Collected OLAP predicates are invalid");
+    TPositionHandle pos = input->Pos;
+
+    auto [pushable, remaining] = SplitForPartialPushdown(predicateTree, false);
+    TVector<TFilterOpsLevels> pushedPredicates;
+    for (const auto& p : pushable) {
+        pushedPredicates.emplace_back(PredicatePushdown(TExprBase(p.ExprNode), lambdaArg, ctx.ExprCtx, pos, pushdownOptions));
+    }
+
+    // TODO: All or nothing currently. Add partial pushdown.
+    if (pushedPredicates.empty() || !remaining.empty()) {
+        return input;
+    }
+
+    const auto& pushedFilter = TFilterOpsLevels::Merge(pushedPredicates, ctx.ExprCtx, pos);
+    const auto remainingFilter = CombinePredicatesWithAnd(remaining, ctx.ExprCtx, pos, false, true);
+
+    TMaybeNode<TExprBase> olapFilter;
+    if (pushedFilter.FirstLevelOps.IsValid()) {
+        // clang-format off
+        olapFilter = Build<TKqpOlapFilter>(ctx.ExprCtx, pos)
+            .Input(lambda.Args().Arg(0))
+            .Condition(pushedFilter.FirstLevelOps.Cast())
+        .Done();
+        // clang-format on
+    }
+
+    if (pushedFilter.SecondLevelOps.IsValid()) {
+        // clang-format off
+        olapFilter = Build<TKqpOlapFilter>(ctx.ExprCtx, pos)
+            .Input(olapFilter.IsValid() ? olapFilter.Cast() : lambda.Args().Arg(0))
+            .Condition(pushedFilter.SecondLevelOps.Cast())
+        .Done();
+        // clang-format on
+    }
+
+    if (!olapFilter.IsValid()) {
+        YQL_CLOG(TRACE, ProviderKqp) << "KqpOlapFilter was not constructed";
+        return input;
+    }
+
+    // clang-format off
+    auto newOlapFilterLambda = Build<TCoLambda>(ctx.ExprCtx, pos)
+        .Args({"olap_filter_row"})
+        .Body<TExprApplier>()
+            .Apply(olapFilter.Cast())
+            .With(lambda.Args().Arg(0), "olap_filter_row")
+            .Build()
+        .Done();
+    // clang-format on
+    YQL_CLOG(TRACE, ProviderKqp) << "Pushed OLAP lambda: " << KqpExprToPrettyString(newOlapFilterLambda, ctx.ExprCtx);
+
+    return std::make_shared<TOpRead>(read->Alias, read->Columns, read->GetOutputIUs(), read->StorageType, read->TableCallable, newOlapFilterLambda.Ptr(),
+                                     read->Pos);
+}
+}
+}
