+++ ydb/core/kqp/opt/rbo/rules/remove_extra_renames_stage.cpp
@@ -1,4 +1,4 @@
-#include "kqp_rbo_rules.h"
+#include <ydb/core/kqp/opt/rbo/kqp_rbo_rules.h>
 #include <ydb/core/kqp/common/kqp_yql.h>
 #include <yql/essentials/core/yql_expr_optimize.h>
 #include <yql/essentials/utils/log/log.h>
